+++ src/main/java/org/dulab/adapcompounddb/models/ontology/Parameters.java
@@ -1,10 +1,10 @@
 package org.dulab.adapcompounddb.models.ontology;
 
 public class Parameters {
-    public static final double MZ_TOLERANCE_PPM = 15.0;
+    public static final int MZ_TOLERANCE_PPM = 15;
     public static final double SCORE_THRESHOLD = 0.3;
-    public static final double PRECURSOR_TOLERANCE_PPM = 5.0;
-    public static final double MASS_TOLERANCE_PPM = 5.0;
+    public static final int PRECURSOR_TOLERANCE_PPM = 5;
+    public static final int MASS_TOLERANCE_PPM = 5;
     public static final double RET_TIME_TOLERANCE = 0.5;
     public static final double ISOTOPIC_SIMILARITY_THRESHOLD = 0.9;
 }
